Fix false function-name linkification in root cause column
Use a stricter regex (_INTERNAL_FUNCTION_PATTERN) that requires
underscores to distinguish pyrefly internal function names like
check_for_imported_final_reassignment() from common Python method
names like get(), match(), set() that appear in error messages.
