+++ scripts/primer_classifier/llm_client.py
@@ -189,7 +189,7 @@ If you have enough context, respond with your analysis. IMPORTANT: reason throug
 The "spec_check", "runtime_behavior", and "mypy_pyright" fields force you to think through the evidence. If the spec doesn't require the check here AND there's no runtime impact AND mypy/pyright wouldn't flag it, it's almost certainly "too strict". The "categories" field is optional — omit it for small diffs with no categories. When present, each entry should correspond to an error category from the input.
 
 PR DIFF ATTRIBUTION:
-When a pyrefly PR diff is provided, identify which specific code change(s) in the diff caused the primer errors to appear or disappear. Reference the file path and function/method from the diff. Be specific — e.g., "the change to `overload_resolution()` in `alt/answers.rs` relaxed the constraint on callable types, which removed the false positive `not-callable` errors." If no PR diff is provided, write "N/A" for pr_attribution.
+When a pyrefly PR diff is provided, identify which specific code change(s) in the diff caused the primer errors to appear or disappear. IMPORTANT: Always use the FULL file path exactly as it appears in the diff (e.g., `pyrefly/lib/alt/answers.rs`, not just `answers.rs`). Reference the function/method name with parentheses (e.g., `overload_resolution()`). Be specific — e.g., "the change to `overload_resolution()` in `pyrefly/lib/alt/answers.rs` relaxed the constraint on callable types, which removed the false positive `not-callable` errors." If no PR diff is provided, write "N/A" for pr_attribution.
 
 In the reason fields, keep explanations concise. When your analysis depends on a typing spec rule, you MUST include the relevant spec link in the reason field (e.g., "per the [typing spec](https://typing.readthedocs.io/en/latest/spec/generics.html#variance), variance inference is only required for protocols")."""
